Restrucutre unit tests and fix error in readme example.
